refactor http client (part 1)
