+++ composer.json
@@ -25,7 +25,7 @@
     "php-http/httplug": "^2.0",
     "php-http/message": "^1.7.1",
     "php-http/message-factory": "^1.0.2",
-    "prooph/event-store": "dev-initial_v8",
+    "prooph/event-store": "dev-master",
     "ramsey/uuid": "^3.8"
   },
   "require-dev": {
