+++ src/Internal/EventStoreHttpConnection.php
@@ -14,7 +14,6 @@ declare(strict_types=1);
 namespace Prooph\EventStoreHttpClient\Internal;
 
 use Http\Message\RequestFactory;
-use Http\Message\UriFactory;
 use Prooph\EventStore\AllEventsSlice;
 use Prooph\EventStore\CatchUpSubscriptionDropped;
 use Prooph\EventStore\CatchUpSubscriptionSettings;
@@ -26,6 +25,7 @@ use Prooph\EventStore\EventAppearedOnCatchupSubscription;
 use Prooph\EventStore\EventAppearedOnPersistentSubscription;
 use Prooph\EventStore\EventAppearedOnSubscription;
 use Prooph\EventStore\EventData;
+use Prooph\EventStore\EventId;
 use Prooph\EventStore\EventReadResult;
 use Prooph\EventStore\EventReadStatus;
 use Prooph\EventStore\EventStoreAllCatchUpSubscription;
@@ -34,63 +34,82 @@ use Prooph\EventStore\EventStorePersistentSubscription;
 use Prooph\EventStore\EventStoreStreamCatchUpSubscription;
 use Prooph\EventStore\EventStoreSubscription;
 use Prooph\EventStore\EventStoreTransaction;
+use Prooph\EventStore\Exception\AccessDeniedException;
+use Prooph\EventStore\Exception\EventStoreConnectionException;
 use Prooph\EventStore\Exception\InvalidArgumentException;
 use Prooph\EventStore\Exception\InvalidOperationException;
 use Prooph\EventStore\Exception\OutOfRangeException;
+use Prooph\EventStore\Exception\StreamDeletedException;
 use Prooph\EventStore\Exception\UnexpectedValueException;
+use Prooph\EventStore\Exception\WrongExpectedVersionException;
 use Prooph\EventStore\ExpectedVersion;
 use Prooph\EventStore\Internal\Consts;
 use Prooph\EventStore\Internal\PersistentSubscriptionCreateResult;
+use Prooph\EventStore\Internal\PersistentSubscriptionCreateStatus;
 use Prooph\EventStore\Internal\PersistentSubscriptionDeleteResult;
+use Prooph\EventStore\Internal\PersistentSubscriptionDeleteStatus;
 use Prooph\EventStore\Internal\PersistentSubscriptionUpdateResult;
+use Prooph\EventStore\Internal\PersistentSubscriptionUpdateStatus;
 use Prooph\EventStore\LiveProcessingStartedOnCatchUpSubscription;
 use Prooph\EventStore\PersistentSubscriptionDropped;
 use Prooph\EventStore\PersistentSubscriptionSettings;
 use Prooph\EventStore\Position;
 use Prooph\EventStore\RawStreamMetadataResult;
+use Prooph\EventStore\ReadDirection;
+use Prooph\EventStore\RecordedEvent;
+use Prooph\EventStore\ResolvedEvent;
+use Prooph\EventStore\SliceReadStatus;
 use Prooph\EventStore\StreamEventsSlice;
 use Prooph\EventStore\StreamMetadata;
 use Prooph\EventStore\StreamMetadataResult;
 use Prooph\EventStore\SubscriptionDropped;
 use Prooph\EventStore\SystemSettings;
 use Prooph\EventStore\UserCredentials;
+use Prooph\EventStore\Util\DateTime;
 use Prooph\EventStore\Util\Json;
 use Prooph\EventStore\WriteResult;
-use Prooph\EventStoreHttpClient\ClientOperations;
 use Prooph\EventStoreHttpClient\ConnectionSettings;
 use Prooph\EventStoreHttpClient\Http\HttpClient;
+use Psr\Http\Client\ClientInterface;
+use Throwable;
 
 /** @internal */
 class EventStoreHttpConnection implements EventStoreConnection
 {
     /** @var HttpClient */
     private $httpClient;
-    /** @var RequestFactory */
-    private $requestFactory;
-    /** @var UriFactory */
-    private $uriFactory;
     /** @var ConnectionSettings */
     private $settings;
+    /** @var callable */
+    private $onException;
     /** @var string */
     private $baseUri;
 
     /** @internal */
     public function __construct(
-        HttpClient $httpClient,
+        ClientInterface $httpClient,
         RequestFactory $requestFactory,
-        UriFactory $uriFactory,
         ConnectionSettings $settings = null
     ) {
-        $this->httpClient = $httpClient;
-        $this->requestFactory = $requestFactory;
-        $this->uriFactory = $uriFactory;
-        $this->settings = $settings ?? ConnectionSettings::default();
         $this->baseUri = \sprintf(
             '%s://%s:%s',
             $this->settings->schema(),
             $this->settings->endPoint()->host(),
             $this->settings->endPoint()->port()
         );
+
+        $this->settings = $settings ?? ConnectionSettings::default();
+
+        $this->httpClient = new HttpClient(
+            $httpClient,
+            $requestFactory,
+            $this->settings,
+            $this->baseUri
+        );
+
+        $this->onException = static function (Throwable $e) {
+            throw new EventStoreConnectionException($e->getMessage());
+        };
     }
 
     public function connectionSettings(): ConnectionSettings
@@ -118,17 +137,39 @@ class EventStoreHttpConnection implements EventStoreConnection
             throw new InvalidArgumentException('Stream cannot be empty');
         }
 
-        return (new ClientOperations\DeleteStreamOperation())(
-            $this->httpClient,
-            $this->requestFactory,
-            $this->uriFactory,
-            $this->baseUri,
-            $stream,
-            $expectedVersion,
-            false,
-            $userCredentials ?? $this->settings->defaultUserCredentials(),
-            $this->settings->requireMaster()
+        $headers = [
+            'ES-ExpectedVersion' => $expectedVersion,
+        ];
+
+        if ($hardDelete) {
+            $headers['ES-HardDelete'] = 'true';
+        }
+
+        if ($this->settings->requireMaster()) {
+            $headers['ES-RequiresMaster'] = 'true';
+        }
+
+        $response = $this->httpClient->delete(
+            '/streams/' . \urlencode($stream),
+            $headers,
+            $userCredentials,
+            $this->onException
         );
+
+        switch ($response->getStatusCode()) {
+            case 204:
+            case 410:
+                return new DeleteResult(Position::invalid());
+            case 400:
+                throw WrongExpectedVersionException::with($stream, $expectedVersion);
+            case 401:
+                throw AccessDeniedException::toStream($stream);
+            default:
+                throw new EventStoreConnectionException(\sprintf(
+                    'Unexpected status code %d returned',
+                    $response->getStatusCode()
+                ));
+        }
     }
 
     /**
@@ -151,17 +192,62 @@ class EventStoreHttpConnection implements EventStoreConnection
             throw new InvalidArgumentException('Stream cannot be empty');
         }
 
-        return (new ClientOperations\AppendToStreamOperation())(
-            $this->httpClient,
-            $this->requestFactory,
-            $this->uriFactory,
-            $this->baseUri,
-            $stream,
-            $expectedVersion,
-            $events,
-            $userCredentials ?? $this->settings->defaultUserCredentials(),
-            $this->settings->requireMaster()
+        $data = [];
+
+        foreach ($events as $event) {
+            \assert($event instanceof EventData);
+
+            $data[] = [
+                'eventId' => $event->eventId()->toString(),
+                'eventType' => $event->eventType(),
+                'data' => $event->data(),
+                'metadata' => $event->metaData(),
+            ];
+        }
+
+        $body = Json::encode($data);
+
+        $headers = [
+            'Content-Type' => 'application/vnd.eventstore.events+json',
+            'Content-Length' => \strlen($body),
+            'ES-ExpectedVersion' => $expectedVersion,
+        ];
+
+        if ($this->settings->requireMaster()) {
+            $headers['ES-RequiresMaster'] = 'true';
+        }
+
+        $response = $this->httpClient->post(
+            '/streams/' . \urlencode($stream),
+            $headers,
+            $body,
+            $userCredentials,
+            $this->onException
         );
+
+        switch ($response->getStatusCode()) {
+            case 201:
+                return new WriteResult(ExpectedVersion::ANY, Position::invalid());
+            case 400:
+                $header = $response->getHeader('ES-CurrentVersion');
+
+                if (empty($header)) {
+                    throw new EventStoreConnectionException($response->getReasonPhrase());
+                }
+
+                $currentVersion = (int) $header[0];
+
+                throw WrongExpectedVersionException::with($stream, $expectedVersion, $currentVersion);
+            case 401:
+                throw AccessDeniedException::toStream($stream);
+            case 410:
+                throw StreamDeletedException::with($stream);
+            default:
+                throw new EventStoreConnectionException(\sprintf(
+                    'Unexpected status code %d returned',
+                    $response->getStatusCode()
+                ));
+        }
     }
 
     public function conditionalAppendToStream(
@@ -191,17 +277,88 @@ class EventStoreHttpConnection implements EventStoreConnection
             throw new OutOfRangeException('Event number is out of range');
         }
 
-        return (new ClientOperations\ReadEventOperation())(
-            $this->httpClient,
-            $this->requestFactory,
-            $this->uriFactory,
-            $this->baseUri,
-            $stream,
-            $eventNumber,
-            $resolveLinkTos,
-            $userCredentials ?? $this->settings->defaultUserCredentials(),
-            $this->settings->requireMaster()
+        $headers = [
+            'Accept' => 'application/vnd.eventstore.atom+json',
+        ];
+
+        if ($this->settings->requireMaster()) {
+            $headers['ES-RequiresMaster'] = 'true';
+        }
+
+        if (-1 === $eventNumber) {
+            $eventNumber = 'head';
+        }
+
+        $response = $this->httpClient->get(
+            '/streams/' . \urlencode($stream) . '/' . $eventNumber . '?embed=tryharder',
+            $headers,
+            $userCredentials,
+            $this->onException
         );
+
+        switch ($response->getStatusCode()) {
+            case 200:
+                $json = Json::decode($response->getBody()->getContents());
+
+                if (empty($json)) {
+                    return new EventReadResult(EventReadStatus::notFound(), $stream, $eventNumber, null);
+                }
+
+                if ($resolveLinkTos && $json['streamId'] !== $stream) {
+                    $data = $json['data'] ?? '';
+
+                    if (\is_array($data)) {
+                        $data = Json::encode($data);
+                    }
+
+                    $field = isset($json['isLinkMetaData']) && $json['isLinkMetaData'] ? 'linkMetaData' : 'metaData';
+
+                    $metadata = $json[$field] ?? '';
+
+                    if (\is_array($metadata)) {
+                        $metadata = Json::encode($metadata);
+                    }
+
+                    $link = new RecordedEvent(
+                        $stream,
+                        $json['positionEventNumber'],
+                        EventId::fromString($json['eventId']),
+                        $json['eventType'],
+                        $json['isJson'],
+                        $data,
+                        $metadata,
+                        DateTime::create($json['updated'])
+                    );
+                } else {
+                    $link = null;
+                }
+
+                $record = new RecordedEvent(
+                    $json['streamId'],
+                    $json['eventNumber'],
+                    EventId::fromString($json['eventId']),
+                    SystemEventTypes::LINK_TO,
+                    false,
+                    $json['title'],
+                    '',
+                    DateTime::create($json['updated'])
+                );
+
+                $event = new ResolvedEvent($record, $link, null);
+
+                return new EventReadResult(EventReadStatus::success(), $stream, $eventNumber, $event);
+            case 401:
+                throw AccessDeniedException::toStream($stream);
+            case 404:
+                return new EventReadResult(EventReadStatus::notFound(), $stream, $eventNumber, null);
+            case 410:
+                return new EventReadResult(EventReadStatus::streamDeleted(), $stream, $eventNumber, null);
+            default:
+                throw new EventStoreConnectionException(\sprintf(
+                    'Unexpected status code %d returned',
+                    $response->getStatusCode()
+                ));
+        }
     }
 
     public function readStreamEventsForward(
@@ -211,37 +368,13 @@ class EventStoreHttpConnection implements EventStoreConnection
         bool $resolveLinkTos = true,
         ?UserCredentials $userCredentials = null
     ): StreamEventsSlice {
-        if (empty($stream)) {
-            throw new InvalidArgumentException('Stream cannot be empty');
-        }
-
-        if ($start < 0) {
-            throw new InvalidArgumentException('Start must be positive');
-        }
-
-        if ($count < 1) {
-            throw new InvalidArgumentException('Count must be positive');
-        }
-
-        if ($count > Consts::MAX_READ_SIZE) {
-            throw new InvalidArgumentException(\sprintf(
-                'Count should be less than %s. For larger reads you should page.',
-                Consts::MAX_READ_SIZE
-            ));
-        }
-
-        return (new ClientOperations\ReadStreamEventsForwardOperation())(
-            $this->httpClient,
-            $this->requestFactory,
-            $this->uriFactory,
-            $this->baseUri,
+        return $this->readStreamEventsForwardPolling(
             $stream,
             $start,
             $count,
             $resolveLinkTos,
             0,
-            $userCredentials ?? $this->settings->defaultUserCredentials(),
-            $this->settings->requireMaster()
+            $userCredentials
         );
     }
 
@@ -272,19 +405,121 @@ class EventStoreHttpConnection implements EventStoreConnection
             ));
         }
 
-        return (new ClientOperations\ReadStreamEventsForwardOperation())(
-            $this->httpClient,
-            $this->requestFactory,
-            $this->uriFactory,
-            $this->baseUri,
-            $stream,
-            $start,
-            $count,
-            $resolveLinkTos,
-            $longPoll,
-            $userCredentials ?? $this->settings->defaultUserCredentials(),
-            $this->settings->requireMaster()
+        $headers = [
+            'Accept' => 'application/vnd.eventstore.atom+json',
+        ];
+
+        if (! $resolveLinkTos) {
+            $headers['ES-ResolveLinkTos'] = 'false';
+        }
+
+        if ($this->settings->requireMaster()) {
+            $headers['ES-RequiresMaster'] = 'true';
+        }
+
+        if ($longPoll > 0) {
+            $headers['ES-LongPoll'] = $longPoll;
+        }
+
+        $response = $this->httpClient->get(
+            '/streams/' . \urlencode($stream) . '/' . $start . '/forward/' . $count . '?embed=tryharder',
+            $headers,
+            $userCredentials,
+            $this->onException
         );
+
+        switch ($response->getStatusCode()) {
+            case 401:
+                throw AccessDeniedException::toStream($stream);
+            case 404:
+                return new StreamEventsSlice(
+                    SliceReadStatus::streamNotFound(),
+                    $stream,
+                    $start,
+                    ReadDirection::forward(),
+                    [],
+                    0,
+                    0,
+                    true
+                );
+            case 410:
+                return new StreamEventsSlice(
+                    SliceReadStatus::streamDeleted(),
+                    $stream,
+                    $start,
+                    ReadDirection::forward(),
+                    [],
+                    0,
+                    0,
+                    true
+                );
+            case 200:
+                $json = Json::decode($response->getBody()->getContents());
+
+                $events = [];
+                $lastEventNumber = 0;
+                foreach (\array_reverse($json['entries']) as $entry) {
+                    if ($json['streamId'] !== $stream) {
+                        $data = $entry['data'] ?? '';
+
+                        if (\is_array($data)) {
+                            $data = Json::encode($data);
+                        }
+
+                        $field = isset($json['isLinkMetaData']) && $json['isLinkMetaData'] ? 'linkMetaData' : 'metaData';
+
+                        $metadata = $json[$field] ?? '';
+
+                        if (\is_array($metadata)) {
+                            $metadata = Json::encode($metadata);
+                        }
+
+                        $link = new RecordedEvent(
+                            $stream,
+                            $json['positionEventNumber'],
+                            EventId::fromString($json['eventId']),
+                            $json['eventType'],
+                            $json['isJson'],
+                            $data,
+                            $metadata,
+                            DateTime::create($json['updated'])
+                        );
+                    } else {
+                        $link = null;
+                    }
+
+                    $record = new RecordedEvent(
+                        $json['streamId'],
+                        $json['eventNumber'],
+                        EventId::fromString($json['eventId']),
+                        SystemEventTypes::LINK_TO,
+                        false,
+                        $json['title'],
+                        '',
+                        DateTime::create($json['updated'])
+                    );
+
+                    $events[] = new ResolvedEvent($record, $link, null);
+
+                    $lastEventNumber = $entry['eventNumber'];
+                }
+
+                return new StreamEventsSlice(
+                    SliceReadStatus::success(),
+                    $stream,
+                    $start,
+                    ReadDirection::forward(),
+                    $events,
+                    $lastEventNumber + 1,
+                    $lastEventNumber,
+                    $json['headOfStream']
+                );
+            default:
+                throw new EventStoreConnectionException(\sprintf(
+                    'Unexpected status code %d returned',
+                    $response->getStatusCode()
+                ));
+        }
     }
 
     public function readStreamEventsBackward(
@@ -309,24 +544,140 @@ class EventStoreHttpConnection implements EventStoreConnection
             ));
         }
 
-        return (new ClientOperations\ReadStreamEventsBackwardOperation())(
-            $this->httpClient,
-            $this->requestFactory,
-            $this->uriFactory,
-            $this->baseUri,
-            $stream,
-            $start,
-            $count,
-            $resolveLinkTos,
-            $userCredentials ?? $this->settings->defaultUserCredentials(),
-            $this->settings->requireMaster()
+        $headers = [
+            'Accept' => 'application/vnd.eventstore.atom+json',
+        ];
+
+        if (! $resolveLinkTos) {
+            $headers['ES-ResolveLinkTos'] = 'false';
+        }
+
+        if ($this->settings->requireMaster()) {
+            $headers['ES-RequiresMaster'] = 'true';
+        }
+
+        $response = $this->httpClient->get(
+            '/streams/' . \urlencode($stream) . '/' . $start . '/backward/' . $count . '?embed=tryharder',
+            $headers,
+            $userCredentials,
+            $this->onException
         );
+
+        switch ($response->getStatusCode()) {
+            case 200:
+                $json = Json::decode($response->getBody()->getContents());
+
+                $events = [];
+                $lastEventNumber = 0;
+                foreach ($json['entries'] as $entry) {
+                    if ($json['streamId'] !== $stream) {
+                        $data = $entry['data'] ?? '';
+
+                        if (\is_array($data)) {
+                            $data = Json::encode($data);
+                        }
+
+                        $field = isset($json['isLinkMetaData']) && $json['isLinkMetaData'] ? 'linkMetaData' : 'metaData';
+
+                        $metadata = $json[$field] ?? '';
+
+                        if (\is_array($metadata)) {
+                            $metadata = Json::encode($metadata);
+                        }
+
+                        $link = new RecordedEvent(
+                            $stream,
+                            $json['positionEventNumber'],
+                            EventId::fromString($json['eventId']),
+                            $json['eventType'],
+                            $json['isJson'],
+                            $data,
+                            $metadata,
+                            DateTime::create($json['updated'])
+                        );
+                    } else {
+                        $link = null;
+                    }
+
+                    $record = new RecordedEvent(
+                        $json['streamId'],
+                        $json['eventNumber'],
+                        EventId::fromString($json['eventId']),
+                        SystemEventTypes::LINK_TO,
+                        false,
+                        $json['title'],
+                        '',
+                        DateTime::create($json['updated'])
+                    );
+
+                    $events[] = new ResolvedEvent($record, $link, null);
+
+                    $lastEventNumber = $entry['eventNumber'];
+                }
+                $nextEventNumber = ($lastEventNumber < 1) ? 0 : ($lastEventNumber - 1);
+
+                return new StreamEventsSlice(
+                    SliceReadStatus::success(),
+                    $stream,
+                    $start,
+                    ReadDirection::backward(),
+                    $events,
+                    $nextEventNumber,
+                    $lastEventNumber,
+                    false
+                );
+            case 401:
+                throw AccessDeniedException::toStream($stream);
+            case 404:
+                return new StreamEventsSlice(
+                    SliceReadStatus::streamNotFound(),
+                    $stream,
+                    $start,
+                    ReadDirection::backward(),
+                    [],
+                    0,
+                    0,
+                    true
+                );
+            case 410:
+                return new StreamEventsSlice(
+                    SliceReadStatus::streamDeleted(),
+                    $stream,
+                    $start,
+                    ReadDirection::backward(),
+                    [],
+                    0,
+                    0,
+                    true
+                );
+            default:
+                throw new EventStoreConnectionException(\sprintf(
+                    'Unexpected status code %d returned',
+                    $response->getStatusCode()
+                ));
+        }
     }
 
     public function readAllEventsForward(
         Position $position,
         int $count,
         bool $resolveLinkTos = true,
+        ?UserCredentials $userCredentials = null
+    ): AllEventsSlice {
+        return $this->readAllEventsForwardPolling(
+            $position,
+            $count,
+            $resolveLinkTos,
+            0,
+            $userCredentials
+        );
+    }
+
+    public function readAllEventsForwardPolling(
+        Position $position,
+        int $count,
+        bool $resolveLinkTos = true,
+        int $longPoll = 0,
         ?UserCredentials $userCredentials = null
     ): AllEventsSlice {
         if ($count < 1) {
@@ -340,17 +691,82 @@ class EventStoreHttpConnection implements EventStoreConnection
             ));
         }
 
-        return (new ClientOperations\ReadAllEventsForwardOperation())(
-            $this->httpClient,
-            $this->requestFactory,
-            $this->uriFactory,
-            $this->baseUri,
-            $position,
-            $count,
-            $resolveLinkTos,
-            $userCredentials ?? $this->settings->defaultUserCredentials(),
-            $this->settings->requireMaster()
+        $headers = [
+            'Accept' => 'application/vnd.eventstore.atom+json',
+        ];
+
+        if (! $resolveLinkTos) {
+            $headers['ES-ResolveLinkTos'] = 'false';
+        }
+
+        if ($this->settings->requireMaster()) {
+            $headers['ES-RequiresMaster'] = 'true';
+        }
+
+        if ($longPoll > 0) {
+            $headers['ES-LongPoll'] = $longPoll;
+        }
+
+        $response = $this->httpClient->get(
+            '/streams/%24all' . '/' . $position->asString() . '/forward/' . $count . '?embed=tryharder',
+            $headers,
+            $userCredentials,
+            $this->onException
         );
+
+        switch ($response->getStatusCode()) {
+            case 401:
+                throw AccessDeniedException::toStream('$all');
+            case 200:
+                $json = Json::decode($response->getBody()->getContents());
+
+                foreach ($json['links'] as $link) {
+                    if ($link['relation'] === 'next') {
+                        $start = \strlen($this->baseUri . '/streams/%24all' . '/');
+                        $nextPosition = Position::parse(\substr($link['uri'], $start, 32));
+                    }
+                }
+
+                $events = [];
+                foreach (\array_reverse($json['entries']) as $entry) {
+                    $data = $entry['data'] ?? '';
+
+                    if (\is_array($data)) {
+                        $data = Json::encode($data);
+                    }
+
+                    $metadata = $json['metadata'] ?? '';
+
+                    if (\is_array($metadata)) {
+                        $metadata = Json::encode($metadata);
+                    }
+
+                    $event = new RecordedEvent(
+                        $json['streamId'],
+                        $json['positionEventNumber'],
+                        EventId::fromString($json['eventId']),
+                        $json['eventType'],
+                        $json['isJson'],
+                        $data,
+                        $metadata,
+                        DateTime::create($json['updated'])
+                    );
+
+                    $events[] = new ResolvedEvent($event, null, null);
+                }
+
+                return new AllEventsSlice(
+                    ReadDirection::forward(),
+                    $position,
+                    $nextPosition,
+                    $events
+                );
+            default:
+                throw new EventStoreConnectionException(\sprintf(
+                    'Unexpected status code %d returned',
+                    $response->getStatusCode()
+                ));
+        }
     }
 
     public function readAllEventsBackward(
@@ -370,17 +786,78 @@ class EventStoreHttpConnection implements EventStoreConnection
             ));
         }
 
-        return (new ClientOperations\ReadAllEventsBackwardOperation())(
-            $this->httpClient,
-            $this->requestFactory,
-            $this->uriFactory,
-            $this->baseUri,
-            $position,
-            $count,
-            $resolveLinkTos,
-            $userCredentials ?? $this->settings->defaultUserCredentials(),
-            $this->settings->requireMaster()
+        $headers = [
+            'Accept' => 'application/vnd.eventstore.atom+json',
+        ];
+
+        if (! $resolveLinkTos) {
+            $headers['ES-ResolveLinkTos'] = 'false';
+        }
+
+        if ($this->settings->requireMaster()) {
+            $headers['ES-RequiresMaster'] = 'true';
+        }
+
+        $response = $this->httpClient->get(
+            '/streams/%24all' . '/' . $position->asString() . '/backward/' . $count . '?embed=tryharder',
+            $headers,
+            $userCredentials,
+            $this->onException
         );
+
+        switch ($response->getStatusCode()) {
+            case 401:
+                throw AccessDeniedException::toStream('$all');
+            case 200:
+                $json = Json::decode($response->getBody()->getContents());
+
+                foreach ($json['links'] as $link) {
+                    if ($link['relation'] === 'previous') {
+                        $start = \strlen($this->baseUri . '/streams/%24all' . '/');
+                        $nextPosition = Position::parse(\substr($link['uri'], $start, 32));
+                    }
+                }
+
+                $events = [];
+                foreach ($json['entries'] as $entry) {
+                    $data = $entry['data'] ?? '';
+
+                    if (\is_array($data)) {
+                        $data = Json::encode($data);
+                    }
+
+                    $metadata = $json['metadata'] ?? '';
+
+                    if (\is_array($metadata)) {
+                        $metadata = Json::encode($metadata);
+                    }
+
+                    $event = new RecordedEvent(
+                        $json['streamId'],
+                        $json['positionEventNumber'],
+                        EventId::fromString($json['eventId']),
+                        $json['eventType'],
+                        $json['isJson'],
+                        $data,
+                        $metadata,
+                        DateTime::create($json['updated'])
+                    );
+
+                    $events[] = new ResolvedEvent($event, null, null);
+                }
+
+                return new AllEventsSlice(
+                    ReadDirection::backward(),
+                    $position,
+                    $nextPosition,
+                    $events
+                );
+            default:
+                throw new EventStoreConnectionException(\sprintf(
+                    'Unexpected status code %d returned',
+                    $response->getStatusCode()
+                ));
+        }
     }
 
     /**
@@ -443,16 +920,11 @@ class EventStoreHttpConnection implements EventStoreConnection
             $metadata
         );
 
-        return (new ClientOperations\AppendToStreamOperation())(
-            $this->httpClient,
-            $this->requestFactory,
-            $this->uriFactory,
-            $this->baseUri,
+        return $this->appendToStream(
             $stream,
             $expectedMetaStreamVersion,
             [$metaEvent],
-            $userCredentials ?? $this->settings->defaultUserCredentials(),
-            $this->settings->requireMaster()
+            $userCredentials
         );
     }
 
@@ -568,17 +1040,41 @@ class EventStoreHttpConnection implements EventStoreConnection
             throw new InvalidArgumentException('Group name cannot be empty');
         }
 
-        return (new ClientOperations\CreatePersistentSubscriptionOperation())(
-            $this->httpClient,
-            $this->requestFactory,
-            $this->uriFactory,
-            $this->baseUri,
-            $stream,
-            $groupName,
-            $settings,
-            $userCredentials ?? $this->settings->defaultUserCredentials(),
-            $this->settings->requireMaster()
+        $body = Json::encode($settings);
+
+        $headers = [
+            'Content-Type' => 'application/json',
+            'Content-Length' => \strlen($body),
+        ];
+
+        if ($this->settings->requireMaster()) {
+            $headers['ES-RequiresMaster'] = 'true';
+        }
+
+        $response = $this->httpClient->put(
+            '/subscriptions/' . \urlencode($stream) . '/' . \urlencode($groupName),
+            $headers,
+            $body,
+            $userCredentials,
+            $this->onException
         );
+
+        $json = Json::decode($response->getBody()->getContents());
+
+        switch ($response->getStatusCode()) {
+            case 401:
+                throw AccessDeniedException::toStream($stream);
+            case 201:
+            case 409:
+                return new PersistentSubscriptionCreateResult(
+                    PersistentSubscriptionCreateStatus::byName($json['result'])
+                );
+            default:
+                throw new EventStoreConnectionException(\sprintf(
+                    'Unexpected status code %d returned',
+                    $response->getStatusCode()
+                ));
+        }
     }
 
     public function updatePersistentSubscription(
@@ -595,17 +1091,42 @@ class EventStoreHttpConnection implements EventStoreConnection
             throw new InvalidArgumentException('Group name cannot be empty');
         }
 
-        return (new ClientOperations\UpdatePersistentSubscriptionOperation())(
-            $this->httpClient,
-            $this->requestFactory,
-            $this->uriFactory,
-            $this->baseUri,
-            $stream,
-            $groupName,
-            $settings,
-            $userCredentials ?? $this->settings->defaultUserCredentials(),
-            $this->settings->requireMaster()
+        $body = Json::encode($settings);
+
+        $headers = [
+            'Content-Type' => 'application/json',
+            'Content-Length' => \strlen($body),
+        ];
+
+        if ($this->settings->requireMaster()) {
+            $headers['ES-RequiresMaster'] = 'true';
+        }
+
+        $response = $this->httpClient->post(
+            '/subscriptions/' . \urlencode($stream) . '/' . \urlencode($groupName),
+            $headers,
+            $body,
+            $userCredentials,
+            $this->onException
         );
+
+        switch ($response->getStatusCode()) {
+            case 200:
+                return new PersistentSubscriptionUpdateResult(
+                    PersistentSubscriptionUpdateStatus::success()
+                );
+            case 401:
+                throw AccessDeniedException::toStream($stream);
+            case 404:
+                return new PersistentSubscriptionUpdateResult(
+                    PersistentSubscriptionUpdateStatus::notFound()
+                );
+            default:
+                throw new EventStoreConnectionException(\sprintf(
+                    'Unexpected status code %d returned',
+                    $response->getStatusCode()
+                ));
+        }
     }
 
     public function deletePersistentSubscription(
@@ -621,16 +1142,32 @@ class EventStoreHttpConnection implements EventStoreConnection
             throw new InvalidArgumentException('Group name cannot be empty');
         }
 
-        return (new ClientOperations\DeletePersistentSubscriptionOperation())(
-            $this->httpClient,
-            $this->requestFactory,
-            $this->uriFactory,
-            $this->baseUri,
-            $stream,
-            $groupName,
-            $userCredentials ?? $this->settings->defaultUserCredentials(),
-            $this->settings->requireMaster()
+        $headers = [];
+
+        if ($this->settings->requireMaster()) {
+            $headers['ES-RequiresMaster'] = 'true';
+        }
+
+        $response = $this->httpClient->delete(
+            '/subscriptions/' . \urlencode($stream) . '/' . \urlencode($groupName),
+            $headers,
+            $userCredentials,
+            $this->onException
         );
+
+        $json = Json::decode($response->getBody()->getContents());
+
+        switch ($response->getStatusCode()) {
+            case 401:
+                throw AccessDeniedException::toStream($stream);
+            case 200:
+            case 404:
+                return new PersistentSubscriptionDeleteResult(
+                    PersistentSubscriptionDeleteStatus::byName($json['result'])
+                );
+            default:
+                throw new \UnexpectedValueException('Unexpected status code ' . $response->getStatusCode() . ' returned');
+        }
     }
 
     public function startTransaction(
